Warrior Cat Game
This was my final project for my very first intro to software class in university. Its a game that allows you to create a cat with stats as if it was the beginning of the video game. It also allows you to save your cat to the computer and then load it back up, so you can save many different cats.